Added movement controller to player ship Changed Integer -> Float in all creator methods Added update actions to DynamicObject Rotated ship.png so it points the same direction as the game obejct

+++ core/src/com/yorkpirates/game/College.java
@@ -2,7 +2,7 @@ package com.yorkpirates.game;
 
 abstract class College extends StaticObject{
 
-    public College(String imgName, Integer xPos, Integer yPos){
+    public College(String imgName, Float xPos, Float yPos){
         super(imgName, xPos, yPos);
     }
