+++ core/src/com/yorkpirates/game/EnemyShip.java
@@ -1,7 +1,7 @@
 package com.yorkpirates.game;
 
 public class EnemyShip extends Ship{
-    public EnemyShip(String imgName, Integer xPos, Integer yPos){
+    public EnemyShip(String imgName, Float xPos, Float yPos){
         super(imgName, xPos, yPos);
     }
