+++ core/src/com/yorkpirates/game/StaticObject.java
@@ -10,14 +10,14 @@ import com.badlogic.gdx.scenes.scene2d.Actor;
 
 abstract class StaticObject extends Actor{
     Sprite sprite;
-    
+
     // init function, sets location and texture
-    public StaticObject(String imgName, Integer xPos, Integer yPos){
+    public StaticObject(String imgName, Float xPos, Float yPos){
         sprite = new Sprite(new Texture(Gdx.files.internal(imgName)));
         setBounds(sprite.getX(), sprite.getY(), sprite.getWidth(), sprite.getHeight()); // v important
         setOrigin(getWidth()/2,getHeight()/2); // Changes center of the object from its (0,0) to the actual center
         setX(xPos);
-        setY(yPos); 
+        setY(yPos);
     }
 
     // Redraws the object every frame, I think.
